+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 
 - Fixed a bug in `Session._write_pandas` where it was erroneously passing `use_logical_type` parameter to `Session._write_modin_pandas_helper` when writing a Snowpark pandas object.
 
+#### Improvements
+
+- Improved the random object name generation to avoid collisions.
+
 #### Experimental Features
 
 - Added support for writing pyarrow Tables to Snowflake tables.